Add HASH_VALUE and HASH_FIND_BY_HASH_VALUE.
This can massively speed up client code that looks up the same key in
multiple hashes, or (for readability reasons) looks up the same key
multiple times in the same hash. If the client code can cache the
HASH_VALUE of the key and reuse it for the multiple lookups, it will
be a massive performance improvement.