Normalize parenthesization of macro arguments.
The rule is "parenthesize all your own arguments, for sanity."
If you're calling another function or macro, you don't need to
parenthesize *their* arguments, because they'll do it. Also,
because your own arguments can't contain unbraced commas, it's
safe to call e.g. `f(a,b)` where `a,b` are your own arguments.
However, it's never safe to use `a->b` without parenthesizing
it: `(a)->b`.

The one intentional exception here is `cmpfcn`, which is allowed
to be a macro name. Invoking `(cmpfcn)(a,b)` would be incorrect
because it would suppress the intended macro expansion of `cmpfcn(a,b)`.