Bump version to 2.0.1 to justify a release tag.
v2.0.1 is API-incompatible with v2.0.0. Yes, this isn't what
semver.org says you ought to do. But I'm considering this a
"bugfix" on top of a "premature release" of v2.0.0; v2.0.0
added new functionality under "incorrect" names and failed
to bump the version number in package.json, both of which
defects are rectified in this release.

v2.0.1 is API-compatible with the v1.9.9.1 release tag, in that
the only API breaks are in features introduced after that tag
was made.