Test case for LL_PREPEND_ELEM, which fails in VS2008 mode.
This is a formal test case for issue #89.
The lack of curly braces around the if/else bodies is intentional,
just to test the exact code that's in tinydtls. That is, if the expansions
of LL_PREPEND_ELEM and LL_APPEND ever lost their do-while-(0) forms,
tinydtls would break; so this tests exactly that case.