Use _utarray_eltptr in place of utarray_eltptr where possible.
This skips the bounds-check, in cases where we know the bounds-check
is redundant.

The one remaining case where we *might* want to use `_utarray_eltptr`
is in `utarray_erase`, when the user is trying to erase e.g. the last
10 elements of a 6-element array. This currently results in correctly
erasing the first 6 elements, and then calling `(a)->icd.dtor` four
times on `NULL`. Arguably it should be undefined behavior to erase a
range that doesn't fall completely inside the array bounds; but the
documentation doesn't say anything one way or the other, so I'm
leaving it alone.