utlist: Add INORDER macros for inserting into sorted lists.
Thanks to @jeffreylovitz for the initial patch!

The `{LL,DL,CDL}_LOWER_BOUND` macros have semantics similar to C++'s
`std::lower_bound` algorithm: in essence they return an iterator to
the first existing list element which is *not less than* the given
one. However, in this context an "iterator to element X" is actually
a "pointer to element X-1", because our lists support only
insertion-after, not insertion-before. Therefore, in situations
where `std::lower_bound` would return `lst.begin()` we return `NULL`,
and in cases where `std::lower_bound` would return `lst.end()` we
return a pointer to the last element of the list.