utlist: Add an assert to silence a scan-build diagnostic.
Thanks to @aaronmdjones for the report!

Unfortunately scan-build is unhappy with uthash and utarray
as well, and the fixes seem harder in those cases. Especially
since there's no tradition of adding redundant asserts in those
files, and I wouldn't really want to start. This case is easy
because we've already committed to liberal use of asserts in
utlist (for whatever reason).