test88's output is HASH_FUNCTION-dependent.
Require that it use HASH_JEN... or at least that it use the
default hash function, whatever that is. If you try to use HASH_BER
or HASH_SFH, you'll get a different pattern of hash collisions,
which means a different number of calls to `alt_memcmp`, which
means different output.

Fixes issue #113.