Add a `uthash_bzero` macro for systems without `memset`.
This just seems like the right thing to do, given that 614b6a5e introduced
similar macros for `memcmp` and `strlen`. I'm not sure why `memset` got a
pass at that time.

Name it `uthash_bzero` because we only ever use it as a standard synonym
for `bzero`. We shouldn't `memset` to any non-zero value, or if we ever do
(and I can't imagine why we would), we should pick a different customization
point for that use-case.