Permit malloc failure to be recoverable via `HASH_NONFATAL_OOM`.
Defining a truthy `HASH_NONFATAL_OOM` now alters the definitions of
the insertion macros so that `uthash_malloc(n)` is allowed to return
`NULL`. When that happens, the insertion will be 100% correctly unwound,
and then we'll call `uthash_nonfatal_oom(elt)` on the element whose
insertion just failed. Because the insertion has already been unwound,
it's 100% safe to `throw` or `longjmp` from `uthash_nonfatal_oom`.
It is also 100% safe to just keep going and check for the non-insertion
back in the caller (by looking for `elt->hh.tbl==NULL`).

By default, if `HASH_NONFATAL_OOM` is not defined, or is defined falsey,
then we use the classic behavior: any `uthash_malloc` failure fails fast
(without unwinding) and hard (by calling `uthash_fatal(msg)`).

Many thanks to @veselov for his painstaking work on this patch!
See #121, #130.

Fixes #92, #93.