hashscan.c: Refactor to resolve GCC `-Wformat-truncation=2` warnings.
The old code took a needlessly baroque approach to formatting each line
of the table. This code is slightly better, and eliminates the need to
compute a maximum size for the `sat` buffer (by eliminating that buffer
entirely).

Also, the block comment explaining the output format had bit-rotted,
so, update that.

`hashscan` itself is still pretty awful and easy to crash, for example
if it thinks it's found a signature in garbage memory for a table with
an unrealistically large `num_buckets` or a key with an unrealistically
large `keylen`. This is not a high priority for me.