Purge `strncpy` from the tests and docs. NFC.
The existing uses of `strncpy` rightly triggered some static
analysis warnings from Coverity, because they were consistently
being used in the pattern

    strncpy(destbuf, src, sizeof destbuf);
    TREAT_AS_STRING(destbuf)

which is wrong: `strncpy` is what you use specifically when there
is a chance that the destination buffer will *not* wind up
null-terminated. In cases like these, we expect (and know) that
the buffer is big enough to store the entire string plus a null
terminator; and so `strcpy` is the correct tool for this job.
(Or arguably `snprintf`, but that's C99 and we still compile
these tests in C89 mode.)

Coverity continues to complain about the `strcpy` itself, in
these tests: `test12.c, test15.c, test66.c, test83.c`.