Rename uthash_memcmp to HASH_KEYCMP, step 1.
Addresses the first half of #157.

We merge this, make a minor release (2.1.0),
and then fix up all the tests (e.g. `test6.c`) to redefine
`HASH_KEYCMP` instead of `uthash_memcmp`.

After a few months like that, we eliminate the mention of
`uthash_memcmp` from `uthash.h` and make a major release (3.0.0).