hashscan: Print the value of num_buckets before trying to malloc it.
If it's a huge number, then the malloc might crash or something.
Let's make sure we print at least some information about it before
then.

Also try to get hashscan compiling on OSX, even though it still won't
run because it depends on the `/proc` filesystem, which is not present
on OSX (it's a Linux thing).

Related to #183.