Remove all support for MurmurHash (HASH_MUR, HASH_USING_NO_STRICT_ALIASING).
This is related to #186: `HASH_MUR` generates a lot of uthash's `-Wcast-align`
noise. There's a lot of crufty technical debt around `HASH_MUR`;
I don't think it's pulling its weight, and will continue to think so
unless somebody steps up to complain about this pull request.

I think that if someone wants to use MurmurHash with uthash, the
right way to do it is to define a function (perhaps an `inline`
function) that implements MurmurHash, and then define `HASH_FUNCTION`
to the name of that function. I don't think we need to continue
supporting MurmurHash out of the box, with all the processor-specific
`#ifdef`s and compiler-specific `-fno-strict-aliasing` cruft that
this old implementation entails.

I've replaced the old test62.c with a new test that shows how to use
`HASH_FUNCTION` and how to use it with `HASH_VALUE` and
`HASH_FIND_BYHASHVALUE`.