unit tests: Fix -Wunused-parameter warnings about argc/argv.
I'm vaguely aware that some embedded compilers insist on the
argc/argv signature (libc++'s unit tests mandate it, for example).
But in this repo we've already got a mix of `int main()` and
`int main(int, char**)`, so clearly nobody's complaining about
`int main()` yet.

Thanks to @obgm for the patch!