Rename uthash_memcmp to HASH_KEYCMP, step 2.
Well, this certainly fell by the wayside! Back in December 2018
(two years ago), I did the first part (introducing `HASH_KEYCMP`),
cut release version 2.1.0, and said we should immediately fix up
the tests and then let it bake for "a few months" before removing
all support for `alt_memcmp`. It turns out that I forgot to do
any of that. So now I've fixed up the tests, just in time to
cut release version 2.2.0.

Then we can let version 2.2.0 bake for "a few months" (read: maybe
a couple of years :)) before finally removing `alt_memcmp`.

test96.c now demonstrates the intended usage of HASH_FUNCTION
and HASH_KEYCMP as a matching pair. (This is similar to the
pair of template parameters `Hasher, KeyEqual` taken by C++'s
`unordered_set`; the difference is that we take `KeyNotEqual`,
so that you can pass in a strcmp-like trivalued comparator --
but you don't have to, and test96.c regression-tests
that passing in a plain old not-equal function works fine.

Fixes #157.