The MCST Elbrus C Compiler supports __typeof. (#247)
The trick here is that MCST's compiler is also colloquially
known as "LCC", so it defines `__LCC__`; so we'd think it was
the [Retargetable C Compiler](https://drh.github.io/lcc/) and
assume it had no `__typeof`. Check for `__MCST__` before checking
for `__LCC__`.

Thanks to @utf-4096 for the patch!