HASH_DEL should be able to delete a const-qualified node
HASH_DEL doesn't need to modify the target node; in fact it should not,
because that target node is usually going straight back to `free` and
any modifications to it would be wasted effort. So, let's make that an
actual tested guarantee, and let's make it const-correct.

Inspired by the discussion in #253.