Follow commenting convention for copyright headers.
Use "/*-" as the first line of copyright header comments to
ease the process of collecting licensing information from a
source tree.

This convention for marking copyright header comments is in
use in other BSD-licensed open-source projects, such as the
FreeBSD and NetBSD operating system projects, elftoolchain,
libarchive, tcpdump/libpcap, and others.

The convention has been formally documented by the FreeBSD
project at [1].

  [1]: https://man.freebsd.org/cgi/man.cgi?query=style

Using this convention in uthash's source tree eases its import
into such downstream projects.  It is a no-op for other uses.

This is a comment-only change that does not affect code
behavior.