Fix MSVC warning C4127 in HASH_BLOOM_TEST (#261)
MSVC `-W4` gives "warning C4127: conditional expression is constant"
for `if (1 != 0)`, but no warning for `if (1)`. Use the latter instead
of the former. In fact, the comparison against zero should really be
all the way down in `HASH_BLOOM_BITTEST`.

No functional change intended.

Thanks to @noodlecollie for the patch!