[test] Respect user's CFLAGS more than our own added ones; remove EXTRA_CFLAGS
See also commit 747bb96, from ten years ago, which first made us
respect the user's CFLAGS here at all. But we still appended our own
flags (e.g. `-I../src -Wall`) to the end of the command line, so the
user couldn't override any of ours with their own.

GNU's guidelines seem like a good way to go:
> https://www.gnu.org/prep/standards/html_node/Command-Variables.html

The EXTRA_CFLAGS variable was introduced to force options to the end
of the command line. Now that the user's CFLAGS are properly at the end,
we can replace our use of EXTRA_CFLAGS with just CFLAGS.

All of this affects only how the tests are compiled, not anything about
the library itself.